% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bips_xaringan.R
\name{bips_xaringan}
\alias{bips_xaringan}
\title{BIPS xaringan presentation format}
\usage{
bips_xaringan(
  css = NULL,
  self_contained = FALSE,
  seal = TRUE,
  yolo = FALSE,
  chakra = "https://remarkjs.com/downloads/remark-latest.min.js",
  nature = list(highlightStyle = "github", highlightLines = TRUE, countIncrementalSlides
    = FALSE, slideNumberFormat = "\%current\%", ratio = "16:9"),
  anchor_sections = FALSE,
  ...
)
}
\arguments{
\item{css}{CSS files to use. Defaults to BIPS theme and xaringan defaults.}

\item{self_contained}{Whether to produce a self-contained HTML file by
embedding all external resources into the HTML file. See the \sQuote{Note}
section below.}

\item{seal}{Whether to generate a title slide automatically using the YAML
metadata of the R Markdown document (if \code{FALSE}, you should write the title
slide by yourself).}

\item{yolo}{Whether to insert the \href{https://kbroman.wordpress.com/2014/08/28/the-mustache-photo/}{Mustache Karl (TM)}
randomly in the slides. \code{TRUE} means insert his picture on one slide, and
if you want him to be on multiple slides, set \code{yolo} to a positive integer
or a percentage (e.g. 0.3 means 30\\% of your slides will be the Mustache
Karl). Alternatively, \code{yolo} can also be a list of the form \code{list(times = n, img = path)}: \code{n} is the number of times to show an image, and \code{path} is
the path to an image (by default, it is Karl).}

\item{chakra}{A path to the remark.js library (can be either local or
remote). Please note that if you use the default remote latest version of
remark.js, your slides will not work when you do not have Internet access.
They might also be broken after a newer version of remark.js is released.
If these issues concern you, you should download remark.js locally (e.g.,
via \code{\link[xaringan:summon_remark]{summon_remark()}}), and use the local version instead.}

\item{nature}{(Nature transformation) A list of configurations to be passed
to \code{remark.create()}, e.g. \code{list(ratio = '16:9', navigation = list(click = TRUE))}; see \url{https://github.com/gnab/remark/wiki/Configuration}. Besides
the options provided by remark.js, you can also set \code{autoplay} to a number
(the number of milliseconds) so the slides will be played every \code{autoplay}
milliseconds; alternatively, \code{autoplay} can be a list of the form
\code{list(interval = N, loop = TRUE)}, so the slides will go to the next page
every \code{N} milliseconds, and optionally go back to the first page to restart
the play when \code{loop = TRUE}. You can also set \code{countdown} to a number (the
number of milliseconds) to include a countdown timer on each slide. If
using \code{autoplay}, you can optionally set \code{countdown} to \code{TRUE} to include a
countdown equal to \code{autoplay}. To alter the set of classes applied to the
title slide, you can optionally set \code{titleSlideClass} to a vector of
classes; the default is \code{c("center", "middle", "inverse")}.}

\item{...}{Additional arguments passed to \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}}.}
}
\description{
Wrapper around \code{\link[xaringan:moon_reader]{xaringan::moon_reader()}} with BIPS styling defaults.
}
